Move enroll command to xpack folder
