+++ x-pack/libbeat/cmd/enroll.go
@@ -1,3 +1,7 @@
+// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+// or more contributor license agreements. Licensed under the Elastic License;
+// you may not use this file except in compliance with the Elastic License.
+
 package cmd
 
 import (
@@ -8,7 +12,7 @@ import (
 
 	"github.com/elastic/beats/libbeat/cmd/instance"
 	"github.com/elastic/beats/libbeat/common/cli"
-	"github.com/elastic/beats/libbeat/management"
+	"github.com/elastic/beats/x-pack/libbeat/management"
 )
 
 func getBeat(name, version string) (*instance.Beat, error) {
