+++ x-pack/libbeat/management/enroll.go
@@ -1,8 +1,12 @@
+// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+// or more contributor license agreements. Licensed under the Elastic License;
+// you may not use this file except in compliance with the Elastic License.
+
 package management
 
 import (
 	"github.com/elastic/beats/libbeat/cmd/instance"
-	"github.com/elastic/beats/libbeat/management/api"
+	"github.com/elastic/beats/x-pack/libbeat/management/api"
 )
 
 // Enroll this beat to the given kibana
