+++ x-pack/winlogbeat/cmd/root.go
@@ -4,11 +4,14 @@
 
 package cmd
 
-import "github.com/elastic/beats/winlogbeat/cmd"
+import (
+	"github.com/elastic/beats/winlogbeat/cmd"
+	xpackcmd "github.com/elastic/beats/x-pack/libbeat/cmd"
+)
 
 // RootCmd to handle beats cli
 var RootCmd = cmd.RootCmd
 
 func init() {
-	// TODO inject x-pack features
+	xpackcmd.AddXPack(RootCmd, cmd.Name)
 }
